testing/web-platform/tests/html/dom/documents/resource-metadata-management/document-lastModified-01.html

<!DOCTYPE html>
<title>document.lastModified should return current local time</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  var d = new Date();
  var localTimeFormatted = ("0" + (d.getHours())).slice(-2)
    + ":" + ("0" + (d.getMinutes())).slice(-2);
  test(function() {
    assert_equals(document.lastModified.substring(11, 16), localTimeFormatted,
     "Hours and minutes should match local time.");
  }, "Date returned by lastModified is in the user's local time zone.");
  var lastMod = document.lastModified;
  var pattern = /[0-9]{2}\/[0-9]{2}\/[0-9]{4} [0-9]{2}:[0-9]{2}:[0-9]{2}/
  var currentDate = new Date();
  var month = ("0" + (currentDate.getMonth() + 1)).slice(-2)
  var day =  ("0" + (currentDate.getDate())).slice(-2)
  var year = currentDate.getFullYear().toString()
  var hours = ("0" + (currentDate.getHours())).slice(-2)
  var minutes = ("0" + (currentDate.getMinutes())).slice(-2)
  var seconds = ("0" + (currentDate.getSeconds())).slice(-2);
  test(function() {
     assert_regexp_match(lastMod, pattern,
       "Format should match the pattern \"NN/NN/NNNN NN:NN:NN\".");
     assert_equals(lastMod.substring(0,2), month, "MM part should match month.");
     assert_equals(lastMod.substring(3,5), day, "DD part should match day.");
     assert_equals(lastMod.substring(6,10), year, "YYYY part should match year.");
     assert_equals(lastMod.substring(11,13), hours, "hh part should match hours.");
     assert_equals(lastMod.substring(14,16), minutes, "mm part should match minutes.");
     assert_approx_equals(parseInt(lastMod.substring(17,19), 10),
       parseInt(seconds), 1, "ss part should match seconds.");
  }, "Date returned by lastModified is in the form \"MM/DD/YYYY hh:mm:ss\".");
  var t = async_test('Date returned by lastModified is current.');
  var old = Date.parse(document.lastModified);
  setTimeout(function() {
    t.step(function() {
      assert_equals((new Date(Date.parse(document.lastModified))).toString(),
        (new Date(old + 1000)).toString());
      t.done();
    });
  }, 1000);
</script>
